+++ src/Components/LoginDialog/logindialog_styles.js
@@ -4,18 +4,30 @@ export const ContainerLogin = styled.div`
     display: flex;
     height: 500px;
     width: 100%;
-    border: 1px solid black;
+.ImageBackground {
+    opacity: 0.4;
+    position: absolute;
+    height: 700px;
+    width: 700px;
+}
+a {
+    text-decoration: none;
+    color: inherit;
 
-    a {
-        text-decoration: none;
-        color: inherit;
-
-        button {
-            height: 30px;
-            width: 200px;
-            margin: 5px;
-        }
+button {
+    height: 30px;
+    width: 200px;
+    margin: 5px;
+    opacity: 1;
+    transition: width 2s, height 2s;
+:hover {
+    color: white;
+    background-color: #48666f;
+    width: 250px;
+    height: 40px;
     }
+}
+}
 `;
 export const FormDiv = styled.div`
     display: flex;
@@ -24,19 +36,41 @@ export const FormDiv = styled.div`
     padding: 40px;
     height: 400px;
     width: 400px;
-    border: 1px solid black;
+    border: 3px solid #71aabb;
     text-align: center;
+    color: #2c3439;
     align-items: center;
-
-    > input {
-        height: 30px;
-        width: 250px;
-        margin: 15px;
-    }
-
-    > button {
-        height: 30px;
-        width: 200px;
-        margin: 5px;
-    }
+    box-shadow: 10px 10px 5px grey;
+    background-color: white;
+    opacity: 0.9;
+    transition: width 2s, height 2s, transform 2s;
+:hover {
+    width: 450px;
+    height: 450px;
+}
+> input {
+    height: 30px;
+    width: 250px;
+    margin: 15px;
+    border: 2px solid #48666f;
+    transition: width 2s, height 2s;
+:hover {
+    width: 300px;
+    height: 35px;
+    border: 2px solid #71aabb;
+}
+}
+> button {
+    height: 30px;
+    width: 200px;
+    margin: 5px;
+    opacity: 1;
+    transition: width 2s, height 2s;
+:hover {
+    color: white;
+    background-color: #48666f;
+    width: 250px;
+    height: 40px;
+}
+}
 `;
